+++ pappl/system.h
@@ -92,7 +92,7 @@ typedef bool (*pappl_save_cb_t)(pappl_system_t *system, void *data);
 // Functions...
 //
 
-extern void		papplSystemAddLink(pappl_system_t *system, const char *label, const char *path_or_url, bool secure) _PAPPL_PUBLIC;
+extern void		papplSystemAddLink(pappl_system_t *system, const char *label, const char *path_or_url, pappl_loptions_t options) _PAPPL_PUBLIC;
 extern bool		papplSystemAddListeners(pappl_system_t *system, const char *name) _PAPPL_PUBLIC;
 extern void		papplSystemAddMIMEFilter(pappl_system_t *system, const char *srctype, const char *dsttype, pappl_mime_filter_cb_t cb, void *data) _PAPPL_PUBLIC;
 extern void		papplSystemAddResourceCallback(pappl_system_t *system, const char *path, const char *format, pappl_resource_cb_t cb, void *data) _PAPPL_PUBLIC;